% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmodeCPRD.R
\name{PlasmodeBin2}
\alias{PlasmodeBin2}
\title{Title Generate Binary Plasmode Simulation Data}
\usage{
PlasmodeBin2(
  formulaOut = NULL,
  objectOut = NULL,
  formulaExp = NULL,
  objectExp = NULL,
  data,
  idVar,
  effectOR = 1,
  MMOut = 1,
  MMExp = 1,
  nsim,
  size,
  eventRate = NULL,
  exposedPrev = NULL
)
}
\arguments{
\item{formulaOut}{The formula to generate outcome. Should be formed like
Outcome ~ Exposure + Confounders}

\item{objectOut}{A fitted model for the outcome model. Can be NULL}

\item{formulaExp}{The formula to generate exposure/treatment.}

\item{objectExp}{The fitted model for treatment model. Can be NUL}

\item{data}{Original data}

\item{idVar}{Name of ID column in data}

\item{effectOR}{User selected effect size}

\item{MMOut}{A multiplier of confounder effects on outcome applied to the
estimated log ORs in the outcome model. Default as 1}

\item{MMExp}{A multiplier of confounder effects on exposure applied to the
estimated log ORs in the exposure model. Default as 1}

\item{nsim}{Number of simulation data}

\item{size}{Size of simulated datasets}

\item{eventRate}{Desired average event rate. Default is the event rate
in the observed data.}

\item{exposedPrev}{Desired average exposure rate. Default is the esposure
prevalence in the observed data.}
}
\description{
Title Generate Binary Plasmode Simulation Data
}
