% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{x_lm_formula}
\alias{x_lm_formula}
\title{Construct Formula for Independent Variables}
\usage{
x_lm_formula(x)
}
\arguments{
\item{x}{A vector of strings representing predictor variable names.}
}
\value{
A single string that concatenates variable names with "+" for use in a model formula.
}
\description{
This function creates a formula string for the predictor variables, with each variable enclosed in backticks and concatenated with "+".
}
\examples{
x_lm_formula(c("age", "weight", "height"))
# Output: "`age` + `weight` + `height`"
}
